+++ lib/transpec/rspec_version.rb
@@ -37,6 +37,10 @@ module Transpec
       @gem_version >= GEM_VERSION_3_0_0_BETA1
     end
 
+    def receive_message_chain_available?
+      @gem_version >= GEM_VERSION_2_99_0_BETA2
+    end
+
     def yielded_example_available?
       @gem_version >= GEM_VERSION_2_99_0_BETA1
     end