+++ spec/transpec/converter_spec.rb
@@ -558,8 +558,8 @@ module Transpec
       let(:method_stub_object) { double('method_stub_object').as_null_object }
 
       shared_examples 'invokes MethodStub#allowize!' do
-        it 'invokes MethodStub#allowize! with #receive_messages availability' do
-          method_stub_object.should_receive(:allowize!).with(rspec_version.receive_messages_available?)
+        it 'invokes MethodStub#allowize! with RSpecVersion' do
+          method_stub_object.should_receive(:allowize!).with(rspec_version)
           converter.process_method_stub(method_stub_object)
         end
       end